+++ crates/factorio-api-gen/src/generate/mod.rs
@@ -13,7 +13,10 @@ mod unions;
 pub use classes::{
     class_names, generate_attribute_setter_lookup, generate_classes, generate_globals,
 };
-pub use concepts::{event_filter_concept_names, generatable_concept_names, generate_concepts};
+pub use concepts::{
+    event_filter_concept_names, flag_set_concept_names, generatable_concept_names,
+    generate_concepts,
+};
 pub use debug_types::generate_debug_types;
 pub use defines::generate_defines;
 pub use event_filters::{generate_event_data, generate_event_filters};