+++ examples/gui_basics/src/lib.rs
@@ -4,11 +4,14 @@
 
 #[factorio_rs::control]
 mod control {
-    use factorio_rs::{factorio_api::classes::LuaGuiElementAddParams, prelude::*};
+    use factorio_rs::{
+        factorio_api::{IndexOrName, classes::LuaGuiElementAddParams},
+        prelude::*,
+    };
 
     #[factorio_rs::event(OnPlayerCreated)]
     pub fn on_player_created(event: OnPlayerCreatedEvent) {
-        if let Some(player) = game.get_player(event.player_index.into()) {
+        if let Some(player) = game.get_player(IndexOrName::Index(event.player_index)) {
             let frame = player.gui().screen().add(LuaGuiElementAddParams {
                 r#type: GuiElementType::Frame,
                 name: Some("gui_basics_root".into()),